move editMediaStreams reducer to its own file, audio and video mute is working
